Fix LSQ unit
